UBER DATA SCRAPPING
This code can take any amount of data off PDFs/CSV/XLSX files or others, it will put all the data, even if its on a PDF invoice/receipt, and puts it on ONE XLSX file. This is my HolyXLSXgrail and feel free to use this code as its great for small businesses/life to make it 1000 times easier to log all your data and you won't have to hire anyone again for this. I made this code with 100 lines at first and managed to condenses it into 11 lines of a SUPER FAST PYTHON CODE, it came straight from my brain, enjoy!